@@ -3,18 +3,18 @@ package com.restaurant.restaurantApi.controller;
 import com.restaurant.restaurantApi.dto.Mensaje;
 import com.restaurant.restaurantApi.model.SettingUser;
 import com.restaurant.restaurantApi.service.inter.ISettingUserService;
-import io.swagger.annotations.Api;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 
 @RestController
 @RequestMapping("setting")
-@Api
+@Tag(name = "Setting User Controller", description = "API para gestión de las credenciales del usuario")
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 public class SettingUserController {
 
