@@ -1,35 +1,20 @@
 package com.restaurant.restaurantApi.common;
 
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-import springfox.bean.validators.configuration.BeanValidatorPluginsConfiguration;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
 
 @Configuration
-@EnableSwagger2
-@Import({BeanValidatorPluginsConfiguration.class})
 public class SwaggerConfig {
 
     @Bean
-    public Docket apiDocket(){
-        return new Docket(DocumentationType.SWAGGER_2)
-                .select()
-                .apis(RequestHandlerSelectors.any())
-                .paths(PathSelectors.any())
-                .build();
+    public OpenAPI customOpenAPI() {
+        return new OpenAPI()
+                .info(new Info()
+                        .title("Restaurant Service API")
+                        .version("1.0")
+                        .description("This API provides all methods required for Restaurant management")
+                );
     }
-
-   /* private ApiInfo getApiInfo(){
-        return new ApiInfo("Restaurant service API", "This API provides all methods required for Account management",
-                "1.0",
-                "TERMS OF SERVICE URL",null,null,null);
-                //new Contact("ELDAR SRL","http://https://www.eldars.com.ar/","humancapital@eldars.com.ar"),
-                //"LICENSE","LICENSE URL", Collections.emptyList());
-    }*/
 }
