@@ -2,8 +2,9 @@ package com.restaurant.restaurantApi.controller;
 
 import com.restaurant.restaurantApi.model.Product;
 import com.restaurant.restaurantApi.service.inter.IProductService;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,7 +14,7 @@ import java.util.List;
 
 @RestController
 @RequestMapping("product")
-@Api
+@Tag(name = "Product Controller", description = "API para gestión de los productos")
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 public class ProductController {
 
@@ -21,7 +22,7 @@ public class ProductController {
     @Autowired
     private IProductService iProductService;
 
-    @ApiOperation(value = "Retorna paginas que tengan esa palabra en su nombre", notes = "Retorna todos los productos que contengan al menos esa letra o palabra en su nombre")
+    @Operation(description = "Retorna paginas que tengan esa palabra en su nombre", method = "Retorna todos los productos que contengan al menos esa letra o palabra en su nombre")
     @GetMapping("getFilterProductByName/{name}")
     public ResponseEntity<List<Product>> filterProductByName(@PathVariable("name") String name){
             return new ResponseEntity<>(this.iProductService.filterProductByName(name),HttpStatus.OK);
