@@ -1,9 +1,9 @@
 package com.restaurant.restaurantApi.dto;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
+
 
 @Data
 public class LoginUsuario {
