@@ -1,9 +1,10 @@
 package com.restaurant.restaurantApi.dto;
 
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
+
 
 @Data
 public class NuevoUsuario {
