@@ -1,14 +1,14 @@
 package com.restaurant.restaurantApi.jwt;
 
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 @Component
@@ -16,9 +16,10 @@ public class JwtEntryPoint  implements AuthenticationEntryPoint {
 
     private final static Logger logger = LoggerFactory.getLogger(JwtEntryPoint.class);
 
+
     @Override
-    public void commence(HttpServletRequest req, HttpServletResponse res, AuthenticationException e) throws IOException, ServletException {
+    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
         logger.error("fail en el método commence");
-        res.sendError(HttpServletResponse.SC_UNAUTHORIZED, "no autorizado");
+        response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "no autorizado");
     }
 }
