@@ -1,6 +1,10 @@
 package com.restaurant.restaurantApi.jwt;
 
-import com.restaurant.restaurantApi.UserDetailsServiceImpl;
+import com.restaurant.restaurantApi.service.impl.UserDetailsServiceImpl;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,10 +13,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.filter.OncePerRequestFilter;
 
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+
 import java.io.IOException;
 
 public class JwtTokenFilter extends OncePerRequestFilter {
@@ -24,6 +25,14 @@ public class JwtTokenFilter extends OncePerRequestFilter {
 
     @Autowired
     UserDetailsServiceImpl userDetailsService;
+
+
+    private String getToken(HttpServletRequest request){
+        String header = request.getHeader("Authorization");
+        if(header != null && header.startsWith("Bearer"))
+            return header.replace("Bearer ", "");
+        return null;
+    }
 
     @Override
     protected void doFilterInternal(HttpServletRequest req, HttpServletResponse res, FilterChain filterChain) throws ServletException, IOException {
@@ -41,12 +50,5 @@ public class JwtTokenFilter extends OncePerRequestFilter {
             logger.error("fail en el método doFilter " + e.getMessage());
         }
         filterChain.doFilter(req, res);
-    }
-
-    private String getToken(HttpServletRequest request){
-        String header = request.getHeader("Authorization");
-        if(header != null && header.startsWith("Bearer"))
-            return header.replace("Bearer ", "");
-        return null;
     }
 }
