@@ -1,10 +1,9 @@
 package com.restaurant.restaurantApi.model;
 
 import com.restaurant.restaurantApi.enums.RolNombre;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
 import lombok.Data;
-
-import javax.persistence.*;
-import javax.validation.constraints.NotNull;
 
 @Entity
 @Data
