@@ -1,14 +1,14 @@
 package com.restaurant.restaurantApi.repo;
 
-import com.restaurant.restaurantApi.model.Usuario;
+import com.restaurant.restaurantApi.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
-    Optional<Usuario> findByNombreUsuario(String nombreUsuario);
+public interface UserRepository extends JpaRepository<User, Integer> {
+    Optional<User> findByNombreUsuario(String nombreUsuario);
     boolean existsByNombreUsuario(String nombreUsuario);
     boolean existsByEmail(String email);
 
