@@ -1,7 +1,7 @@
-package com.restaurant.restaurantApi;
+package com.restaurant.restaurantApi.service.impl;
 
-import com.restaurant.restaurantApi.model.Usuario;
-import com.restaurant.restaurantApi.model.UsuarioPrincipal;
+import com.restaurant.restaurantApi.model.User;
+import com.restaurant.restaurantApi.model.PrimaryUser;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -11,12 +11,12 @@ import org.springframework.stereotype.Service;
 @Service("userDetailsService")
 public class UserDetailsServiceImpl implements UserDetailsService {
     @Autowired
-    private UsuarioService usuarioService;
+    private UsuarioServiceImpl usuarioServiceImpl;
 
     @Override
     public UserDetails loadUserByUsername(String nombreUsuario) throws UsernameNotFoundException {
-        Usuario usuario = usuarioService.getByNombreUsuario(nombreUsuario).get();
-        return UsuarioPrincipal.build(usuario);
+        User user = usuarioServiceImpl.getByNombreUsuario(nombreUsuario).get();
+        return PrimaryUser.build(user);
     }
 
 
