@@ -1,12 +1,12 @@
-package com.restaurant.restaurantApi;
+package com.restaurant.restaurantApi.service.inter;
 
 import com.restaurant.restaurantApi.model.Rol;
 import com.restaurant.restaurantApi.enums.RolNombre;
 import com.restaurant.restaurantApi.repo.RolRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import javax.transaction.Transactional;
 import java.util.Optional;
 
 @Service
