@@ -1,9 +1,8 @@
 package com.restaurant.restaurantApi.service.impl;
 
-import com.restaurant.restaurantApi.UsuarioService;
 import com.restaurant.restaurantApi.enums.RolNombre;
 import com.restaurant.restaurantApi.model.Rol;
-import com.restaurant.restaurantApi.model.Usuario;
+import com.restaurant.restaurantApi.model.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,62 +20,62 @@ import static org.junit.jupiter.api.Assertions.*;
 public class AuthServiceTest {
 
     @Autowired
-    UsuarioService usuarioService;
+    UsuarioServiceImpl usuarioServiceImpl;
 
-    private Usuario usuario;
+    private User user;
 
 
     @BeforeEach
     public void set_up(){
         Rol rol = new Rol(RolNombre.ROLE_ADMIN);
-        usuario = new Usuario();
-        usuario.setNombre("user");
-        usuario.setNombreUsuario("user123");
-        usuario.setPassword("123");
-        usuario.setEmail("user@gmail.com");
-        usuario.setRol(rol);
-        this.usuarioService.save(usuario);
+        user = new User();
+        user.setNombre("user");
+        user.setNombreUsuario("user123");
+        user.setPassword("123");
+        user.setEmail("user@gmail.com");
+        user.setRol(rol);
+        this.usuarioServiceImpl.save(user);
     }
 
     @Test
     public void guardar_usuario(){
         Rol rol = new Rol(RolNombre.ROLE_ADMIN);
-        Usuario user = new Usuario();
-        user = new Usuario();
+        User user = new User();
+        user = new User();
         user.setNombre("luigi");
         user.setNombreUsuario("luigi123");
         user.setPassword("123");
         user.setEmail("luigi@gmail.com");
         user.setRol(rol);
-        Usuario usuarioGuardado = this.usuarioService.save(user);
-        assertNotNull(usuarioGuardado.getId());
-        assertEquals(usuarioGuardado.getNombre(),"luigi");
-        assertEquals(usuarioGuardado.getNombreUsuario(),"luigi123");
-        assertEquals(usuarioGuardado.getPassword(),"123");
-        assertEquals(usuarioGuardado.getEmail(),"luigi@gmail.com");
-        assertEquals(usuarioGuardado.getRol().getRolNombre(),RolNombre.ROLE_ADMIN);
+        User userGuardado = this.usuarioServiceImpl.save(user);
+        assertNotNull(userGuardado.getId());
+        assertEquals(userGuardado.getNombre(),"luigi");
+        assertEquals(userGuardado.getNombreUsuario(),"luigi123");
+        assertEquals(userGuardado.getPassword(),"123");
+        assertEquals(userGuardado.getEmail(),"luigi@gmail.com");
+        assertEquals(userGuardado.getRol().getRolNombre(),RolNombre.ROLE_ADMIN);
     }
     @Test
     public void existe_mail(){
-        Boolean existsEmail = this.usuarioService.existsByEmail("user@gmail.com");
+        Boolean existsEmail = this.usuarioServiceImpl.existsByEmail("user@gmail.com");
         assertTrue(existsEmail);
     }
 
     @Test
     public void existe_nombre_usuario(){
-        Boolean existsNombreUsuario = this.usuarioService.existsByNombreUsuario("user123");
+        Boolean existsNombreUsuario = this.usuarioServiceImpl.existsByNombreUsuario("user123");
         assertTrue(existsNombreUsuario);
     }
 
     @Test
     public void obtener_usuario_por_nombre_de_usuario(){
-        Usuario usuario = this.usuarioService.getByNombreUsuario("user123").get();
-        assertNotNull(usuario.getId());
-        assertEquals(usuario.getNombre(),"user");
-        assertEquals(usuario.getNombreUsuario(),"user123");
-        assertEquals(usuario.getEmail(),"user@gmail.com");
-        assertEquals(usuario.getPassword(),"123");
-        assertEquals(usuario.getRol().getRolNombre(),RolNombre.ROLE_ADMIN);
+        User user = this.usuarioServiceImpl.getByNombreUsuario("user123").get();
+        assertNotNull(user.getId());
+        assertEquals(user.getNombre(),"user");
+        assertEquals(user.getNombreUsuario(),"user123");
+        assertEquals(user.getEmail(),"user@gmail.com");
+        assertEquals(user.getPassword(),"123");
+        assertEquals(user.getRol().getRolNombre(),RolNombre.ROLE_ADMIN);
     }
 
 }
